# Prevent _events creation on emit()