--- a/lib/eventemitter2.js
+++ b/lib/eventemitter2.js
@@ -298,6 +298,9 @@
   };
 
   EventEmitter.prototype.emit = function() {
+    if (!this._events && !this._all) {
+      return false;
+    }
 
     this._events || init.call(this);
 
@@ -404,6 +407,9 @@
   };
 
   EventEmitter.prototype.emitAsync = function() {
+    if (!this._events && !this._all) {
+      return false;
+    }
 
     this._events || init.call(this);