@RangerMauve 2018-10-16
Hi! I don't have time to do a full review right now, but I'll to get on this by next Monday.

Would you mind pinging me on here if I don't get it done by then?

I'm thinking this should be sent out as a breaking change in case somebody out there is relying on the old behavior.