@RangerMauve 2019-12-23
Thanks for the PR. Published it in 6.0.0